# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/application/code": {
            "post": {
                "description": "用户点击\"获取验证码\"按钮，系统向用户提供的邮箱发送6位验证码，用户需要在申请表单中填入验证码才可以成功完成身份验证，否则不应该可以提交申请。验证码时限为10分钟，超时无效",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "获取申请验证码 Vera",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.GetVerifyCodeQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\": \"邮件发送成功\",\"status\": 200}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"数据格式错误\", \"status\": 400}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"msg\": \"没有该用户\", \"status\": 401}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "{\"msg\": \"验证码存储失败\",\"status\": 402}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "{\"msg\": \"发送邮件失败\",\"status\": 403}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/application/create": {
            "post": {
                "description": "用户可以申请认领自己的学者门户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "申请学者门户 Vera \u0026 txc",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.CreateApplicationQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\": \"申请成功\", \"status\": 200,\"application\": application}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"数据格式错误\", \"status\": 400}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"msg\": \"没有该用户\", \"status\": 401}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "{\"msg\": \"创建申请失败\", \"status\": 402}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "{\"msg\": \"该学者已被认领\", \"status\": 403}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/application/handle": {
            "post": {
                "description": "管理员对用户提交的申请进行审核，并给出审核意见content",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "审核学者门户申请 Vera \u0026 txc",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.HandleApplicationQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\": \"审核成功\", \"status\": 200}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"数据格式错误\", \"status\": 400}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "{\"msg\": \"管理员不存在 or 不是管理员\", \"status\": 401}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "{\"msg\": \"申请不存在\", \"status\": 402}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "{\"msg\": \"已审核过该申请\", \"status\": 403}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "{\"msg\": \"数据库错误\", \"status\": 404}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.CreateApplicationQ": {
            "type": "object",
            "required": [
                "author_id",
                "author_name",
                "email",
                "institution",
                "user_id",
                "verify_code"
            ],
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "author_name": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "institution": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "verify_code": {
                    "type": "string"
                }
            }
        },
...
        "response.GetVerifyCodeQ": {
            "type": "object",
            "required": [
                "email",
                "user_id"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "response.HandleApplicationQ": {
            "type": "object",
            "required": [
                "application_id",
                "status",
                "user_id"
            ],
            "properties": {
                "application_id": {
                    "type": "integer"
                },
                "content": {
                    "description": "审批意见",
                    "type": "string"
                },
                "status": {
                    "description": "是否通过",
                    "type": "integer"
                },
                "user_id": {
                    "description": "谁审核的",
                    "type": "integer"
                }
            }
        },
...
    }
}